# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Opaque record payload (the full client-side record), or null for tombstones.
 * @nullable
 */
export type SyncPayload = { [key: string]: unknown } | null;

export interface SyncChange {
  kind: string;
  recordId: string;
  payload?: SyncPayload | null;
  /** Epoch milliseconds, monotonic per device. */
  clientUpdatedAt: number;
  deleted: boolean;
}

export interface SyncRecord {
  kind: string;
  recordId: string;
  payload?: SyncPayload | null;
  clientUpdatedAt: number;
  deleted: boolean;
  revision: number;
}

export interface SyncInput {
  changes: SyncChange[];
  /** Server revision cursor. Use "0" for a full pull. */
  cursor: string;
  /** Optional pull page size. */
  limit?: number;
}

export interface SyncResult {
  records: SyncRecord[];
  cursor: string;
  hasMore: boolean;
  /** Server clock (epoch ms) at response time. */
  serverTime: number;
}