# lib/api-zod/src/generated/types/syncChange.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncPayload } from "./syncPayload";

export interface SyncChange {
  kind: string;
  recordId: string;
  payload?: SyncPayload | null;
  /** Epoch milliseconds, monotonic per device. */
  clientUpdatedAt: number;
  deleted: boolean;
}