# lib/api-zod/src/generated/types/syncInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncChange } from "./syncChange";

export interface SyncInput {
  changes: SyncChange[];
  /** Server revision cursor. Use "0" for a full pull. */
  cursor: string;
  /** Optional pull page size. */
  limit?: number;
}