# lib/api-zod/src/generated/types/syncPayload.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Opaque record payload (the full client-side record), or null for tombstones.
 * @nullable
 */
export type SyncPayload = { [key: string]: unknown } | null;