# lib/api-zod/src/generated/types/syncRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncPayload } from "./syncPayload";

export interface SyncRecord {
  kind: string;
  recordId: string;
  payload?: SyncPayload | null;
  clientUpdatedAt: number;
  deleted: boolean;
  revision: number;
}