# lib/api-zod/src/generated/types/syncResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SyncRecord } from "./syncRecord";

export interface SyncResult {
  records: SyncRecord[];
  cursor: string;
  hasMore: boolean;
  /** Server clock (epoch ms) at response time. */
  serverTime: number;
}